This commit was generated by cvs2svn to compensate for changes in r217, which included commits to RCS files with non-trunk default branches.

+++ src/libafs/redhat.sh
@@ -5,15 +5,20 @@
 IBYTES=""
 SETATTR=""
 
-grep i_bytes /lib/modules/$1/build/include/linux/fs.h > /dev/null
+if [ -e $1/include/linux/fs.h ] ; then
+grep i_bytes $1/include/linux/fs.h > /dev/null
 if [ $? = 0 ]; then
-    IBYTES="-DSTRUCT_INODE_HAS_I_BYTES=1"
-fi
-grep "extern int inode_setattr" /lib/modules/$1/build/include/linux/fs.h > /dev/null
-if [ $? = 0 ]; then
-    SETATTR="-DINODE_SETATTR_NOT_VOID=1"
+IBYTES="-DSTRUCT_INODE_HAS_I_BYTES=1"
 fi
 
+grep "extern int inode_setattr" $1/include/linux/fs.h > /dev/null
+if [ $? = 0 ]; then
+SETATTR="-DINODE_SETATTR_NOT_VOID=1"
+fi
+
+if [ -e $2 ] ; then
 /bin/rm $2
+fi
 echo "KDEFINES = ${IBYTES} ${SETATTR}" > $2
+fi
 exit 0
